+++ TransientTest/test/TransientUse.t.sol
@@ -17,21 +17,21 @@ contract TransientUseTest is Test {
     function testReentrancyProtection() public {
         vm.expectRevert(bytes("Attack failed"));
         badActor.callbackFunction();
-     }
+    }
 
     function testMultiplierInSingleTransaction() public {
         transientUse.setMultiplier(3);
-        uint result = transientUse.multiply(5);
+        uint256 result = transientUse.multiply(5);
         assertEq(result, 15, "Multiplication result should be 15");
-    } 
-      
+    }
+
     function testMultiplierAcrossTransactions() public {
         transientUse.setMultiplier(3);
-        
+
         // Simulate a new transaction by rolling to next block
         vm.roll(block.number + 1);
-        
-        uint result = transientUse.multiply(5);
+
+        uint256 result = transientUse.multiply(5);
         assertEq(result, 0, "Multiplication result should be 0 after new transaction");
     }
 }
